fix: replace all operations with Stripe to Stripe Utility Service
