+++ payment-service/src/main/java/com/nikonenko/paymentservice/controllers/PaymentCustomerController.java
@@ -1,13 +1,16 @@
 package com.nikonenko.paymentservice.controllers;
 
+import com.nikonenko.paymentservice.dto.customers.CustomerRideRequest;
+import com.nikonenko.paymentservice.dto.customers.CustomerRideResponse;
 import com.nikonenko.paymentservice.dto.customers.StripeCustomerChargeRequest;
 import com.nikonenko.paymentservice.dto.customers.StripeCustomerChargeResponse;
 import com.nikonenko.paymentservice.dto.customers.StripeCustomerRequest;
 import com.nikonenko.paymentservice.dto.customers.StripeCustomerResponse;
-import com.nikonenko.paymentservice.services.StripeCustomerServiceImpl;
+import com.nikonenko.paymentservice.services.PaymentCustomerService;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -17,18 +20,24 @@ import org.springframework.web.bind.annotation.RestController;
 @RestController
 @RequiredArgsConstructor
 @RequestMapping("api/v1/stripe/customer")
-public class StripeCustomerController {
-    private final StripeCustomerServiceImpl stripeCustomerService;
+public class PaymentCustomerController {
+    private final PaymentCustomerService paymentCustomerService;
 
     @PostMapping()
     @ResponseStatus(HttpStatus.CREATED)
-    public StripeCustomerResponse createCustomer(@RequestBody @Valid StripeCustomerRequest customerRequest){
-        return stripeCustomerService.createCustomer(customerRequest);
+    public StripeCustomerResponse createCustomer(@RequestBody @Valid StripeCustomerRequest customerRequest) {
+        return paymentCustomerService.createCustomer(customerRequest);
+    }
+
+    @GetMapping("/ride-price")
+    @ResponseStatus(HttpStatus.OK)
+    public CustomerRideResponse calculateRidePrice(@RequestBody @Valid CustomerRideRequest customerRideRequest) {
+        return paymentCustomerService.calculateRidePrice(customerRideRequest);
     }
 
     @PostMapping("/charge")
     @ResponseStatus(HttpStatus.OK)
     public StripeCustomerChargeResponse customerCharge(@RequestBody @Valid StripeCustomerChargeRequest request) {
-        return stripeCustomerService.customerCharge(request);
+        return paymentCustomerService.customerCharge(request);
     }
 }
